requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
pdf2image
google-cloud-aiplatform